Introduce strategy pattern for conversions
