+++ src/otoole/preprocess/__init__.py
@@ -1,19 +1,13 @@
 from .excel_to_osemosys import generate_csv_from_excel
 from .create_datapackage import main as create_datapackage, csv_to_datapackage
-from .datapackage_to import (
-    convert_datapackage_to_datafile,
-    convert_datapackage_to_excel,
-)
 from .datafile_to_datapackage import convert_file_to_package, read_datafile_to_dict
 
 
 __all__ = [
     "generate_csv_from_excel",
     "create_datapackage",
-    "convert_datapackage_to_datafile",
     "convert_file_to_package",
     "create_datapackage_from_datafile",
-    "convert_datapackage_to_excel",
     "read_datafile_to_dict",
     "csv_to_datapackage",
 ]
